Backend Connected
Successfully implimented the backend for our website. Login and Signup page both work. Users can create an account as long as the email ends in @columbus.edu. After logging in, login and signup buttons are disabled and a logout button is shown. Lastly, we have begun to design the Profile page and successfully created the Contact page.