Item List Implementation
In this update, we have implemented the item list that is retrieved from the database. Currently, images for the items are not implemented, but will be in the near future. In addition to this, a sell item page is in its initial development, and will be completed on the next development session.